refactor(module): master, slave db 생성 후 불필요한 메서드 정리
