+++ src/main/java/nextstep/subway/line/application/LineService.java
@@ -14,7 +14,6 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import javax.persistence.Temporal;
 import java.util.List;
 import java.util.stream.Collectors;
 
@@ -33,7 +32,7 @@ public class LineService {
             @CacheEvict(cacheNames = "paths", allEntries = true),
             @CacheEvict(cacheNames = "lines", allEntries = true)
     })
-    @Transactional()
+    @Transactional
     public LineResponse saveLine(LineRequest request) {
         Station upStation = stationService.findById(request.getUpStationId());
         Station downStation = stationService.findById(request.getDownStationId());
