+++ src/main/java/nextstep/subway/line/domain/Line.java
@@ -191,14 +191,4 @@ public class Line extends BaseEntity implements Serializable {
             .filter(it -> it.equalUpStation(stationId))
             .findFirst();
     }
-
-    @Override
-    public String toString() {
-        return "Line{" +
-                "id=" + id +
-                ", name='" + name + '\'' +
-                ", color='" + color + '\'' +
-                ", sections=" + sections +
-                '}';
-    }
 }
