+++ src/main/java/nextstep/subway/station/application/StationService.java
@@ -23,7 +23,6 @@ public class StationService {
         this.stationRepository = stationRepository;
     }
 
-
     @CacheEvict(cacheNames = "stations", allEntries = true)
     @Transactional
     public StationResponse saveStation(StationRequest stationRequest) {
@@ -32,7 +31,6 @@ public class StationService {
     }
 
     @Cacheable(cacheNames = "stations", unless = "#result.isEmpty()")
-    @Transactional(readOnly = true)
     public List<StationResponse> findAllStations(Long id, Pageable pageable) {
         List<Station> stations = stationRepository.findAll(id, pageable);
