+++ src/main/java/nextstep/subway/station/domain/StationRepository.java
@@ -8,6 +8,7 @@ import org.springframework.data.repository.query.Param;
 import java.util.List;
 
 public interface StationRepository extends JpaRepository<Station, Long> {
+
     @Query("SELECT s FROM Station s WHERE s.id >= :id")
     List<Station> findAll(@Param("id") Long id, Pageable pageable);
 }
